[Test] Globally mock sentry and telemetry
